refactor nodeSync client
